TootTally v0.4

+++ Utils/TootTallyLogger.cs
@@ -3,10 +3,6 @@ using BepInEx.Logging;
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using static Mono.Security.X509.X520;
 
 namespace TootTally.Utils
 {
@@ -14,11 +10,13 @@ namespace TootTally.Utils
     {
         private const string TOOTTALLY_LOG_FOLDER = "Logs";
         private const string TOOTTALLY_LOG_FILE_NAME = "TootTally.log";
+        private static List<string> _initializedLogs;
 
         public static void Initialize()
         {
             var folderPath = Path.Combine(Paths.BepInExRootPath, TOOTTALLY_LOG_FOLDER);
             var logFilePath = Path.Combine(folderPath, TOOTTALLY_LOG_FILE_NAME);
+            _initializedLogs = new List<string>();
             if (!Directory.Exists(folderPath))
             {
                 LogInfo("Couldn't find logs folder, generating folder...");
@@ -43,6 +41,12 @@ namespace TootTally.Utils
             Plugin.GetLogger().LogError(msg);
         }
 
+        internal static void CatchError(Exception ex)
+        {
+            Plugin.GetLogger().LogError(ex.Message);
+            Plugin.GetLogger().LogError(ex.StackTrace);
+        }
+
         internal static void LogWarning(string msg)
         {
             Plugin.GetLogger().LogWarning(msg);
@@ -61,47 +65,42 @@ namespace TootTally.Utils
 
         private static void OnLogEvent(object sender, LogEventArgs e)
         {
-            var filePath = Path.Combine(Paths.BepInExRootPath, TOOTTALLY_LOG_FOLDER, TOOTTALLY_LOG_FILE_NAME);
-            if (!File.Exists(filePath))
-                File.Create(filePath);
-            var level = e.Level.ToString();
-            var source = e.Source.SourceName;
-            if (source == "TootTally")
-                File.AppendAllText(filePath, $"[{DateTime.Now:HH:mm:ss}]   {level,-8}[Core] {e.Data}\n");
-            else
-                File.AppendAllText(filePath, $"[{DateTime.Now:HH:mm:ss}]   {level,-8}[{source.Remove(0, 10)}] {e.Data}\n");
-            if ((sender as ManualLogSource) != Plugin.GetLogger())
+            try
             {
-                var sourceFilePath = Path.Combine(Paths.BepInExRootPath, TOOTTALLY_LOG_FOLDER, e.Source.SourceName + ".log");
-                File.AppendAllText(sourceFilePath, $"[{DateTime.Now:HH:mm:ss}]   {level,-8}[{source.Remove(0, 10)}] {e.Data}\n");
+                var filePath = Path.Combine(Paths.BepInExRootPath, TOOTTALLY_LOG_FOLDER, TOOTTALLY_LOG_FILE_NAME);
+                if (!File.Exists(filePath))
+                    File.Create(filePath).Close();
+                var level = e.Level.ToString();
+                var source = e.Source.SourceName;
+                if (source == "TootTally")
+                    File.AppendAllText(filePath, $"[{DateTime.Now:HH:mm:ss}]   {level,-8}[Core] {e.Data}\n");
+                else
+                    File.AppendAllText(filePath, $"[{DateTime.Now:HH:mm:ss}]   {level,-8}[{source.Remove(0, 10)}] {e.Data}\n");
+                if ((sender as ManualLogSource) != Plugin.GetLogger())
+                {
+                    var sourceFilePath = Path.Combine(Paths.BepInExRootPath, TOOTTALLY_LOG_FOLDER, e.Source.SourceName + ".log");
+                    File.AppendAllText(sourceFilePath, $"[{DateTime.Now:HH:mm:ss}]   {level,-8}[{source.Remove(0, 10)}] {e.Data}\n");
+                }
             }
-        }
+            catch (Exception ex)
+            {
+                Plugin.GetLogger().LogError("TootTally Logger couldn't write the log.");
+                Plugin.GetLogger().LogError(ex.Message);
+                Plugin.GetLogger().LogError(ex.StackTrace);
+            }
 
-        public static void LogInfo(ManualLogSource logger, string msg)
-        {
-            logger.LogInfo(msg);
-        }
-        public static void DebugModeLog(ManualLogSource logger, string msg)
-        {
-            if (Plugin.Instance.DebugMode.Value)
-                logger.LogDebug(msg);
-        }
-        public static void LogError(ManualLogSource logger, string msg)
-        {
-            logger.LogError(msg);
-        }
-
-        public static void LogWarning(ManualLogSource logger, string msg)
-        {
-            logger.LogWarning(msg);
         }
 
         public static void ClearOrCreateLogFile(string logFileName)
         {
             var sourceFilePath = Path.Combine(Paths.BepInExRootPath, TOOTTALLY_LOG_FOLDER, logFileName + ".log");
-            if (File.Exists(sourceFilePath))
-                File.Delete(sourceFilePath);
-            File.Create(sourceFilePath);
+            if (!_initializedLogs.Contains(sourceFilePath))
+            {
+                if (File.Exists(sourceFilePath))
+                    File.Delete(sourceFilePath);
+                File.Create(sourceFilePath).Close();
+                _initializedLogs.Add(sourceFilePath);
+            }
         }
 
     }
